Extract address formatting to a concern
