+++ app/services/notify/renewal_letter_service.rb
@@ -4,8 +4,7 @@ require "notifications/client"
 
 module Notify
   class RenewalLetterService < ::WasteCarriersEngine::BaseService
-    # So we can use displayable_address()
-    include ::WasteCarriersEngine::ApplicationHelper
+    include CanFormatAddress
 
     def run(registration:)
       @registration = NotifyRenewalLetterPresenter.new(registration)
@@ -14,24 +13,6 @@ module Notify
 
       client.send_letter(template_id: template,
                          personalisation: personalisation)
-    end
-
-    private
-
-    def address_lines
-      address_values = [
-        @registration.contact_name,
-        displayable_address(@registration.contact_address)
-      ].flatten
-
-      address_hash = {}
-
-      address_values.each_with_index do |value, index|
-        line_number = index + 1
-        address_hash["address_line_#{line_number}"] = value
-      end
-
-      address_hash
     end
   end
 end
